housekeeping: Fixing frontend scaffolding
