--- tools/scaffolding/templates/frontend/jest.config.js
@@ -1,8 +1,0 @@
-const baseConfig = require("@clutch-sh/tools/jest.config");
-
-module.exports = {
-  ...baseConfig,
-  coverageThreshold: {
-    ...baseConfig.coverageThreshold,
-  },
-};
