+++ tools/scaffolding/templates/frontend/workflow/external/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@clutch-sh/{{ .PackageName }}",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "description": " {{ .Description }}",
   "license": "Apache-2.0",
@@ -22,24 +22,36 @@
     {{- if .IsWizardTemplate}}
     "@clutch-sh/wizard": "^4.0.0-beta",
     {{- end}}
+    "esbuild": "^0.18.0",
+    "eslint": "^8.3.0",
+    "jest": "^27.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-is": "^17.0.2"
+    "typescript": "^4.2.3"
   },
   "resolutions": {
+    "@types/node": "18.19.23",
     "@types/react": "17.0.50",
+    "@types/react-dom": "17.0.17",
     "@mui/material": "5.8.5",
     "@mui/lab": "5.0.0-alpha.87",
     "@mui/styles": "5.8.4",
     "@mui/system": "5.8.5",
-    "react-hook-form": "7.25.3"
+    "esbuild": "0.18.13",
+    "eslint": "8.16.0",
+    "jest": "27.5.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-hook-form": "7.25.3",
+    "react-router": "6.0.0-beta.0",
+    "react-router-dom": "6.0.0-beta.0",
+    "typescript": "4.2.3"
   },
   "devDependencies": {
     "@clutch-sh/tools": "^4.0.0-beta"
   },
   "engines": {
-    "node": ">=18 <19",
-    "yarn": "^4.3.1"
-  },
-  "packageManager": "yarn@4.3.1"
+    "node": ">=18",
+    "yarn": "^4.0.0"
+  }
 }
