+++ tools/scaffolding/templates/frontend/workflow/external/tsconfig.json
@@ -1,19 +1,30 @@
 {
   "compilerOptions": {
+    "allowSyntheticDefaultImports": true,
+    "baseUrl": "./",
+    "composite": true,
     "declaration": true,
+    "declarationMap": true,
     "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
     "importsNotUsedAsValues": "error",
     "incremental": true,
+    "isolatedModules": true,
     "jsx": "react",
     "lib": ["DOM", "ESNext"],
+    "module": "esnext",
     "moduleResolution": "node",
     "noImplicitAny": false,
     "noImplicitReturns": true,
     "outDir": "./dist",
     "preserveConstEnums": true,
+    "preserveSymlinks": true,
+    "resolveJsonModule": true,
     "rootDir": "./src",
+    "skipLibCheck": true,
+    "strict": false,
     "target": "ES2018",
-    "types": ["jest", "node", "long"],
+    "types": ["jest", "node", "long"]
   },
   "include": [
     "src/**/*",
