+++ tools/scaffolding/templates/frontend/workflow/internal/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require("@clutch-sh/tools/.eslintrc");
