+++ tools/scaffolding/templates/frontend/workflow/internal/babel.config.js
