+++ tools/scaffolding/templates/frontend/workflow/internal/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@clutch-sh/tools/jest.config");
