+++ tools/scaffolding/templates/frontend/workflow/internal/prettier.config.js
