+++ tools/scaffolding/templates/frontend/workflow/internal/src/hello-world.tsx
