+++ tools/scaffolding/templates/frontend/workflow/internal/src/index.tsx
